refactor: 🚚 Déplaçage du fichier decoration dans les designs
